+++ examples/example_project/lib/receiver.ex
@@ -34,38 +34,38 @@ defmodule Example.Receiver do
 
   @impl true
   def handle_notification({:stream_id, stream_id} = msg, _from, _ctx, state) do
-    Membrane.Logger.debug(msg)
+    Membrane.Logger.debug("#{inspect(msg)}")
     state = Map.put(state, :stream_id, stream_id)
     {{:ok, forward: {:source, {:gather_candidates, stream_id}}}, state}
   end
 
   @impl true
   def handle_notification({:new_candidate_full, _candidate} = msg, _from, _ctx, state) do
-    Membrane.Logger.debug(msg)
+    Membrane.Logger.debug("#{inspect(msg)}")
     {:ok, state}
   end
 
   @impl true
   def handle_notification(:gathering_done = msg, _from, _ctx, state) do
-    Membrane.Logger.debug(msg)
+    Membrane.Logger.debug("#{inspect(msg)}")
     {{:ok, forward: {:source, {:get_local_credentials, state.stream_id}}}, state}
   end
 
   @impl true
   def handle_notification({:local_credentials, _credentials} = msg, _from, _ctx, state) do
-    Membrane.Logger.debug(msg)
+    Membrane.Logger.debug("#{inspect(msg)}")
     {:ok, state}
   end
 
   @impl true
   def handle_other({:set_remote_credentials, remote_credentials, stream_id} = msg, _ctx, state) do
-    Membrane.Logger.debug(msg)
+    Membrane.Logger.debug("#{inspect(msg)}")
     {{:ok, forward: {:source, {:set_remote_credentials, remote_credentials, stream_id}}}, state}
   end
 
   @impl true
   def handle_other({:set_remote_candidate, candidate} = msg, _ctx, state) do
-    Membrane.Logger.debug(msg)
+    Membrane.Logger.debug("#{inspect(msg)}")
     {{:ok, forward: {:source, {:set_remote_candidate, candidate, state.stream_id, 1}}}, state}
   end
 end
