+++ lib/membrane_ice/source.ex
@@ -1,13 +1,27 @@
 defmodule Membrane.ICE.Source do
   use Membrane.Source
-  use Membrane.ICE.Common
+
+  require Unifex.CNode
 
   alias Membrane.Buffer
+  alias Membrane.ICE.Common
 
   def_output_pad :output,
     availability: :always,
     caps: :any,
     mode: :push
+
+  @impl true
+  def handle_init(_options) do
+    {:ok, cnode} = Unifex.CNode.start_link(:native)
+    :ok = Unifex.CNode.call(cnode, :init)
+
+    state = %{
+      cnode: cnode
+    }
+
+    {:ok, state}
+  end
 
   @impl true
   def handle_other(
@@ -27,5 +41,10 @@ defmodule Membrane.ICE.Source do
     metadata = %{:stream_id => stream_id, :component_id => component_id}
     actions = [buffer: {:output, %Buffer{payload: payload, metadata: metadata}}]
     {{:ok, actions}, state}
+  end
+
+  @impl true
+  def handle_other(msg, context, state) do
+    Common.handle_ice_message(msg, context, state)
   end
 end
