+++ mix.exs
@@ -37,11 +37,14 @@ defmodule Membrane.ICE.Mixfile do
 
   defp deps do
     [
-      {:membrane_core, git: "https://github.com/membraneframework/membrane_core.git", branch: "master", override: true},
-#      {:unifex,
-#       git: "https://github.com/membraneframework/unifex.git",
-#       branch: "implement-cnode-string",
-#       override: true},
+      {:membrane_core,
+       git: "https://github.com/membraneframework/membrane_core.git",
+       branch: "master",
+       override: true},
+      #      {:unifex,
+      #       git: "https://github.com/membraneframework/unifex.git",
+      #       branch: "implement-cnode-string",
+      #       override: true},
       {:unifex, path: "/home/michal/Repos/unifex", override: true},
       {:ex_doc, "~> 0.22", only: :dev, runtime: false},
       {:dialyxir, "~> 1.0.0", only: :dev, runtime: false}
